Abstract card out into it's own component
