+++ src/app/components/CardsContainer.tsx
@@ -1,6 +1,6 @@
 'use client'
 import React, { MouseEvent, useState } from 'react';
-import { Card } from './Card';
+import { UserCard } from './UserCard';
 
 export interface User {
   id: string;
@@ -31,7 +31,6 @@ export const CardsContainer = ({ userData }: CardsContainerProps) => {
   const [activeCard, setActiveCard] = useState<number | undefined>(undefined);
 
   const handleCardClick = (index: number) => {
-    console.log(index)
     setActiveCard(index)
   }
 
@@ -39,9 +38,9 @@ export const CardsContainer = ({ userData }: CardsContainerProps) => {
     <div className="grid grid-cols-2 w-full gap-4">
         {userData.map((user: User, index: number) => {
           return (
-            <Card key={index} {...formatUserData(user)} isActive={index === activeCard} onClick={(e) => handleCardClick(index)}>
+            <UserCard key={index} {...formatUserData(user)} isActive={index === activeCard} onClick={(e) => handleCardClick(index)}>
               <img src={user.image} />
-            </Card>
+            </UserCard>
           )    
         })}
       </div>
